requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
torch==2.6.0
torchaudio==2.6.0
librosa==0.11.0
numpy==1.26.4
scipy==1.15.3
pyyaml==6.0.3
modelscope==1.34.0
transformers==4.41.2
requests==2.32.5
tqdm==4.67.1
zhconv==1.4.3
zhon==2.1.1
jiwer==4.0.0
faster-whisper==1.2.1
funasr==1.2.9
pymcd==0.2.1
praat-parselmouth==0.4.7
soundfile==0.13.1
sheet-sqa==0.2.5
s3prl==0.4.10
addict==2.4.0
datasets==3.6.0
pillow==12.1.0